refactor(wallet): extract useWalletInit and useWalletBalance hooks
